[monarch][debugger] Split debugger.py into smaller files and add `monarch debug` to existing CLI
This diff makes no substantive changes. It just splits up `debugger.py` into multiple smaller files and moves the `monarch.debug_cli` module's functionality into the pre-existing `monarch` CLI tool. So now you access the debugger by running `monarch debug` instead of `python -m monarch.debug_cli`.

Differential Revision: [D81983068](https://our.internmc.facebook.com/intern/diff/D81983068/)

[ghstack-poisoned]
