--- python/monarch/debug_cli/__main__.py
@@ -1,12 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-# pyre-unsafe
-from monarch._src.debug_cli import debug_cli
-
-
-if __name__ == "__main__":
-    debug_cli.run()
