+++ python/monarch/tools/cli.py
@@ -9,10 +9,16 @@ import argparse
 import json
 import sys
 
+from monarch._src.actor.debugger.env import (
+    _get_debug_server_host,
+    _get_debug_server_port,
+)
+
 from monarch.tools.commands import (
     bounce,
     component_args_from_cli,
     create,
+    debug,
     info,
     kill,
     stop,
@@ -141,6 +147,25 @@ class StopCmd:
         stop(args.server_handle)
 
 
+class DebugCmd:
+    def add_arguments(self, subparser: argparse.ArgumentParser) -> None:
+        subparser.add_argument(
+            "--host",
+            type=str,
+            default=_get_debug_server_host(),
+            help="Hostname where the debug server is running",
+        )
+        subparser.add_argument(
+            "--port",
+            type=int,
+            default=_get_debug_server_port(),
+            help="Port that the debug server is listening on",
+        )
+
+    def run(self, args: argparse.Namespace) -> None:
+        debug(args.host, args.port)
+
+
 def get_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser(description="Monarch CLI")
     subparser = parser.add_subparsers(title="COMMANDS")
@@ -149,6 +174,7 @@ def get_parser() -> argparse.ArgumentParser:
         "create": CreateCmd(),
         "info": InfoCmd(),
         "kill": KillCmd(),
+        "debug": DebugCmd(),
         # --- placeholder subcommands (not yet implemented) ---
         "bounce": BounceCmd(),
         "stop": StopCmd(),
@@ -162,6 +188,9 @@ def get_parser() -> argparse.ArgumentParser:
 def main(argv: list[str] = sys.argv[1:]) -> None:
     parser = get_parser()
     args = parser.parse_args(argv)
+    if not hasattr(args, "func"):
+        parser.print_help()
+        sys.exit(1)
     args.func(args)
